% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_racmo_in_space.R
\name{average_racmo_in_space}
\alias{average_racmo_in_space}
\title{Calculate the mean value of a RACMO variable over a given extent}
\usage{
average_racmo_in_space(
  racmoData,
  extent,
  minArea = 0,
  weight = FALSE,
  extentArgs = list()
)
}
\arguments{
\item{racmoData}{SpatRaster: The RACMO data to average. It must be an
existing SpatRaster.}

\item{extent}{Define the spatial extent used to average over. Fed directly
into \code{get_extent()}; see there for details. See also 'extentArgs'.}

\item{minArea}{numeric: Threshold value to determine the minimum fraction
of a pixel that must be covered by the 'extent' to be included in the
calculation. Pixels with a lower fraction are simply excluded from the
calculation (Step 2 in the Description).}

\item{weight}{BINARY: Should each pixels' values be weighted by the pixels'
area in the 'extent'? Example: if a pixel has a value of 4 kg m-2, and is
75 \% covered by the 'extent', should the value be 4 (FALSE) or 3 (TRUE)?}

\item{extentArgs}{list: Any arguments that need to be passed to
\code{get_extent()} to define the 'extent'. It is probably easier to define
'extent' elsewhere and then just feed it into this function rather
than define this list.}
}
\description{
Calculates the average value of a RACMO variable across a
given extent. A single value is returned for each layer of the input
'racmoData'. Optionally, pixels can be discarded if the pixel's area
within the 'extent' is too low (set with 'minArea'). Pixels can also be
weighted by their area within the 'extent'.

\if{html}{\out{<div class="sourceCode">}}\preformatted{The function simply:
   1) extracts the value at each pixel
   2) discards any unnecessary pixels (i.e. area in extent < minArea)
   3) optionally, weights the pixel values with the pixel's extent area
   4) calculates the mean across the remaining values
   5) returns a dataframe containing the average value for each layer,
      and the corresponding date of the layer
}\if{html}{\out{</div>}}
}
\examples{
-----------------------------------------------------------------
\dontrun{
  # racmoData
  x <- subset_racmoM_by_summer("precip", 1991:1994) |>
  subset_racmoM_by_month(c(11, 12, 1, 2))

  # shelf extent
  y <- get_shelf_outline("Shackleton")

  # Average value across Shackleton each month
  z <- average_racmo_in_space(racmoData = x, extent = y, minArea = 0.5)
  print(z)
}

}
