% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_multimonth_racmo.R
\name{calc_austral_racmo}
\alias{calc_austral_racmo}
\title{Calculate RACMO values over multiple months each austral summer / year}
\usage{
calc_austral_racmo(racmoData, months, FUN, australSplit = 3, ...)
}
\arguments{
\item{racmoData}{SpatRaster: The RACMO data to use with 'FUN'. It must be
an existing SpatRaster.}

\item{months}{vector: Which month/s to include? Input can be the month
number (e.g. 12) or the month name, either in full ("December",
"december") or abbreviated ("Dec", "dec"). Multiple months can be input
at once (e.g. c(12, 1, 2)), but do not try to mix strings and numbers in
the vector.}

\item{FUN}{Which function should be applied? Examples include "mean", "sd",
and "median". This function is applied to each pixel across the months
each year. For example, "mean" gives each pixel's average value in the
months each year; using "sum" would calculate the total across those
months.}

\item{australSplit}{numeric: Which is the last month included in an austral
summer before the new austral year begins? The default value is 3, which
means that all months \emph{AFTER} March are considered as part of the
following summer (i.e. April 1991 -- March 1992 are all in 1992). Swap
this value according: setting it as 4 means May 1991 -- April 1992 are
all 1992.}

\item{...}{Any arguments that should be passed to the 'FUN' function.}
}
\description{
Often, it is necessary to look at RACMO values across multiple
months, but across austral summers / years. For example, what is the
total DJF precipitation each austral year / summer? Or what is the
average wind speed in DJF? This function takes the input RACMO data, and
creates new SpatRasters based on the requested 'months' and the 'FUN'
(function) arguments.

The following example further illustrates this function. Take a monthly
RACMO dataset of average monthly wind speeds, including all months each
year from 1979--2018 (i.e. 12 months * 40 years = 480 layers). Set
'months' as c(12, 1, 2) (i.e. DJF), and FUN as "mean". The returned
SpatRaster will contain 40 layers, with each one being that austral
summers's average DJF wind speed. The date of each returned layer is the
date of the first of the constituent layers that it was computed from.
Only layers with the same units are processed together, and only austral
years when all requested months are available are returned.
}
\seealso{
calc_annual_racmo
}
