% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fun_in_timeM.R
\name{calc_fun_in_timeM}
\alias{calc_fun_in_timeM}
\title{Calculate RCM values across multiple months each year / austral summer}
\usage{
calc_fun_in_timeM(x, months, annual, FUN, ...)
}
\arguments{
\item{x}{SpatRaster: The RACMO or MAR data to use with 'FUN'. It must be
an existing SpatRaster.}

\item{months}{vector: Which month/s to include? Input can be the month
number (e.g. 12) or the month name, either in full ("December",
"december") or abbreviated ("Dec", "dec"). Multiple months can be input
at once (e.g. c(12, 1, 2)), but do not try to mix strings and numbers in
the vector.}

\item{annual}{BINARY: Should the months be split based on calendar years
(TRUE), or an austral summer? If the latter, provide a numeric value
between 1 and 12 to indicate which is the last month included in an
austral summer before the new austral year begins. For example, a value
of 3 means that all months \emph{AFTER} March are considered as part of the
following year / summer (i.e. April 1991 -- March 1992 are all in 1992).}

\item{FUN}{Which function should be applied? Examples include "mean", "sd",
and "median". This function is applied to each pixel across the months
each year. For example, "mean" gives each pixel's average value across
the months each year; using "sum" would calculate the total across those
months each year.}

\item{...}{Any arguments that should be passed to the 'FUN' function.}
}
\description{
Often, it is necessary to look at RCM values across multiple
months each year / summer. For example, what is the total JJA
precipitation each year? Or what is the average wind speed in DJF each
summer? This function takes the input RCM data, and creates new
SpatRasters based on the requested 'months' and the 'FUN' (function)
arguments, accounting for whether the months should be split into
calendar years or austral summers / years.

The following examples further illustrates this function.
\subsection{Example 1 - Annual Winter (JJA) total precipitation}{

Take a monthly RACMO dataset of precipitation, including all months
each year from 1979--2018 (i.e. 12 months * 40 years = 480 layers). Set
'months' as c(6, 7, 8) (i.e. JJA), and FUN as "sum". The returned
SpatRaster will contain 40 layers, with each one being that year's total
JJA precipitation. The date of each returned layer is the date of the
first of the constituent layers that it was computed from. Only layers
with the same units are processed together, and only years when all
requested months are available are returned.
}

\subsection{Example 2 - Austral Summer (DJF) mean wind speed}{

Take a monthly MAR dataset of average monthly wind speeds, including
all months each year from 1979--2018 (i.e. 12 months * 40 years = 480
layers). Set 'months' as c(12, 1, 2) (i.e. DJF), FUN as "mean" and annual
as 3. The returned SpatRaster will contain 40 layers, with each one being
that summers's average DJF wind speed. The date of each returned layer is
the date of the first of constituent layers that it was computed from.
Only layers with the same units are processed together, and only austral
summers / years when all requested months are available are returned.
}
}
