% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_multimonth_racmo.R
\name{calc_multiMonth_racmo}
\alias{calc_multiMonth_racmo}
\title{Calculate RACMO values over multiple months each year / austral summer}
\usage{
calc_multiMonth_racmo(racmoData, months, annual, FUN, ...)
}
\arguments{
\item{racmoData}{SpatRaster: The RACMO data to use with 'FUN'. It must be
an existing SpatRaster.}

\item{months}{vector: Which month/s to include? Input can be the month
number (e.g. 12) or the month name, either in full ("December",
"december") or abbreviated ("Dec", "dec"). Multiple months can be input
at once (e.g. c(12, 1, 2)), but do not try to mix strings and numbers in
the vector.}

\item{annual}{BINARY: Should the months be split based on calendar years
(TRUE), or an austral summer? If the latter, provide a numeric value
between 1 and 12 to indicate which is the last month included in an
austral summer before the new austral year begins? For example, a value
of 3 means that all months \emph{AFTER} March are considered as part of the
following summer (i.e. April 1991 -- March 1992 are all in 1992).}

\item{FUN}{Which function should be applied? Examples include "mean", "sd",
and "median". This function is applied to each pixel across the months
each year. For example, "mean" gives each pixel's average value in the
months each year; using "sum" would calculate the total across those
months.}

\item{...}{Any arguments that should be passed to the 'FUN' function.}
}
\description{
Often, it is necessary to look at RACMO values across multiple
months each year / summer. For example, what is the total JJA
precipitation each year? Or what is the average wind speed in DJF each
summer? This function takes the input RACMO data, and creates new
SpatRasters based on the requested 'months' and the 'FUN' (function)
arguments, accounting for whether the months should be split into
calendar years or austral summer years.

The following examples further illustrates this function.
\enumerate{
\item Take a monthly RACMO dataset of precipitation, including all months
each year from 1979--2018 (i.e. 12 months * 40 years = 480 layers). Set
'months' as c(6, 7, 8) (i.e. JJA), and FUN as "sum". The returned
SpatRaster will contain 40 layers, with each one being that year's total
JJA precipitation. The date of each returned layer is the date of the
first of the constituent layers that it was computed from. Only layers
with the same units are processed together, and only years when all
requested months are available are returned.
\item Take a monthly RACMO dataset of average monthly wind speeds, including
all months each year from 1979--2018 (i.e. 12 months * 40 years = 480
layers). Set 'months' as c(12, 1, 2) (i.e. DJF), and FUN as "mean". The
returned SpatRaster will contain 40 layers, with each one being that
summers's average DJF wind speed. The date of each returned layer is the
date of the first of constituent layers that it was computed from. Only
layers with the same units are processed together, and only austral years
when all requested months are available are returned.
}
}
