% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_racmoD_climatology.R
\name{calc_racmoD_climatology}
\alias{calc_racmoD_climatology}
\title{Apply a function to daily RACMO data across multiple years}
\usage{
calc_racmoD_climatology(racmoData, FUN, excludeIncomplete = FALSE, ...)
}
\arguments{
\item{racmoData}{SpatRaster: The daily RACMO data. It needs to be an
existing SpatRaster.}

\item{FUN}{Which function/s should be applied across the dataset? For
example, if set as "mean", this function will return the mean value of
the racmoData variable on the 1st Dec each year. Can be a vector of
functions, but be wary of speed as there is no elegant solution to
supplying '...' for different functions; it is probably easier to run
this function separately for each required 'FUN' function and then
combine them as a SpatRasterDataset.}

\item{excludeIncomplete}{Should data be excluded from the function if the
day is not found throughout the dataset? TRUE removes dates not found in
every calendar year (i.e. Jan--Dec); a numeric value between 1 and 12
does the same but based on an austral year, and this argument defines the
'australSplit' (e.g. if 3, all months AFTER March, are considered part of
the following austral year / summer). If FALSE (the default), all data is
used when calculating the function.

Can be useful so the function is applied to the same years/summers for
each day, but be careful! Worst case scenario: if a dataset ran from 2nd
January 2000 to 1st January 2001, no data would be used because none of
the dates occur in both 2000 and 2001. See
\code{terrapin::exclude_unmatched_days()}. It is likely much easier and
clearer to establish which dates are in the dataset first, and then run
this function with 'excludeIncomplete' set as FALSE.}

\item{...}{Any arguments that need to be fed to the 'FUN' functions. If
there are multiple values in the 'FUN' vector (i.e. multiple functions
will be applied), the "..." arguments will be fed to all of the
functions.}
}
\description{
What is the mean air temperature on the 1st December each
summer? What is the maximum wind speed on the days from the 10th to the
15th July each year? What is the standard deviation in surface melt on
the days in the first weeks of January? This function can help answer
such questions.
}
