% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_racmo.R
\name{crop_racmo_to_basin}
\alias{crop_racmo_to_basin}
\title{Crop RACMO data to an Antarctic ice shelf}
\usage{
crop_racmo_to_basin(racmoData, basin)
}
\arguments{
\item{racmoData}{SpatRaster: The RACMO data to crop. It must be an existing
SpatRaster.}

\item{basin}{"string": The EXACT name of the basins to crop the RACMO
data to. If a vector with multiple basins is input, non-shelf pixels in
the bounding box between those basin are also included.}
}
\description{
Spatially mask RACMO SpatRasters. 'basin' must be an exact
match for a name in the MEaSURES or IMBIE dataset. Multiple basins can be
used in a vector. The data returned is for a bounding box around the
shelves; use \code{mask_racmo_to_basin()} for "cropping" tightly to the actual
basin outline.
}
\examples{
-----------------------------------------------------------------
  dontrun{
    x <- subset_racmoM_by_summer("precip", 1991) |>
      subset_racmoM_by_month(12)

    # Single basin
    xCrop <- crop_racmo_to_basin(x, "A-Ap")
    terra::plot(xCrop)

    # Multiple basins
    xCrop2 <- crop_racmo_to_basin(x, c("A-Ap", "Dry Valleys"))
    terra::plot(xCrop2)
  }

}
\seealso{
mask_racmo_to_basin()

crop_racmo_to_shelf()
}
