% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_racmo.R
\name{crop_racmo_to_shelf}
\alias{crop_racmo_to_shelf}
\title{Crop RACMO data to an Antarctic ice shelf}
\usage{
crop_racmo_to_shelf(racmoData, shelf)
}
\arguments{
\item{racmoData}{SpatRaster: The RACMO data to crop. It must be an existing
SpatRaster.}

\item{shelf}{"string": The EXACT name of the ice shelves to crop the RACMO
data to. If a vector with multiple shelves is input, non-shelf pixels in
the bounding box between those shelves are also included.}
}
\description{
Spatially crop RACMO SpatRasters. 'shelf' must be an exact
match for a name in the MEaSURES dataset. Multiple shelves can be used in
a vector. The data returned is for a bounding box around the shelves; use
\code{mask_racmo_to_shelf()} for "cropping" tightly to the actual shelf
outline.
}
\examples{
-----------------------------------------------------------------
  \dontrun{
    x <- subset_racmoM_by_summer("precip", 1991) |>
      subset_racmoM_by_month(12)

    # Single shelf
    xCrop <- crop_racmo_to_shelf(x, "Amery")
    terra::plot(xCrop)

    # Multiple shelves
    xCrop2 <- crop_racmo_to_shelf(x, c("Amery", "Shackleton"))
    terra::plot(xCrop2)
  }

}
\seealso{
mask_racmo_to_shelf()

crop_racmo_to_basin()
}
