% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_racmo.R
\name{crop_racmoD}
\alias{crop_racmoD}
\title{Crop daily RACMO data to a spatial extent}
\usage{
crop_racmoD(racmoData, extent, rectangularExtent = TRUE, ...)
}
\arguments{
\item{racmoData}{SpatRaster: The RACMO data to crop. Can be either a
variable name, in which case raw daily RACMO data is read in; or an
existing SpatRaster of daily RACMO data.}

\item{extent}{The area to crop the RACMO data to; fed directly into
\code{racmoR::get_extent()} so see there for details.}

\item{rectangularExtent}{BINARY: Should the RACMO data be cropped to a
rectangular bounding box around the extent (TRUE), or to the actual
outline (e.g. shelf/basin) of the extent (FALSE)? The latter is achieved
by masking out pixels (setting their values as NA) if they do not contain
part of the extent outline.}

\item{...}{Any other arguments that can be fed into \code{get_extent()}.}
}
\description{
This function is a simple wrapper around
\code{racmoR::crop_racmo()} and \code{racmoR::read_racmoD_data()} to spatially crop
daily RACMO SpatRasters.
}
