% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_to_x.R
\name{crop_to_shelf}
\alias{crop_to_shelf}
\title{Crop RCM data to an Antarctic ice shelf}
\usage{
crop_to_shelf(x, shelf)
}
\arguments{
\item{x}{SpatRaster: The RACMO or MAR data to crop. It must be an existing
SpatRaster.}

\item{shelf}{"string": The EXACT name of the ice shelves to crop 'x' to. If
a vector with multiple shelves is input, non-shelf pixels in the bounding
box between those shelves are also included.}
}
\description{
Spatially crop RACMO or MAR SpatRasters. The 'shelf' must be
an exact match for a name in the MEaSURES dataset. Multiple shelves can
be used in a vector. The data returned is for a bounding box around the
shelves; use \code{mask_to_shelf()} for "cropping" tightly to the actual shelf
outline.
}
\examples{
-----------------------------------------------------------------
\dontrun{
  x <- subset_racmoM_by_summer("precip", 1991) |>
    subset_racmoM_by_month(12)

  # Single shelf
  xCrop <- crop_to_shelf(x, "Amery")
  terra::plot(xCrop)

  # Multiple shelves
  xCrop2 <- crop_to_shelf(x, c("Amery", "Shackleton"))
  terra::plot(xCrop2)
}

}
\seealso{
mask_to_shelf()

crop_to_basin()
}
