% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_racmo_globals.R
\name{define_racmo_globals}
\alias{define_racmo_globals}
\title{Define and provide access to global RACMO datasets and variables}
\usage{
define_racmo_globals(
  rawDataPath = "../../Data/",
  racmoM = "RACMO/RACMO2.3p3_CON_ANT27_monthly/",
  racmoD = "RACMO/RACMO2.3p3_CON_ANT27_daily/",
  MEaSURES = "MEaSURES Boundaries/"
)
}
\arguments{
\item{rawDataPath}{"string": The relative path to the folder where raw data
is stored. The approach taken here assumes that all raw data is held in a
"Data" directory containing subfolders for the "RACMO/racmoM",
"RACMO/racmoD" and "MEaSURES" datasets. If this is not the case, amend
this function as necessary.}

\item{racmoM}{"string": The name of the subfolder containing the raw
monthly RACMO NetCDF files. Set as NULL if no daily data is necessary /
available; if so, "racmoM" functions will not be available.}

\item{racmoD}{"string": The name of the subfolder containing the raw daily
RACMO NetCDF files. Set as NULL if no daily data is necessary /
available; if so, "racmoD" functions will not be available.}

\item{MEaSURES}{"string": The name of the subfolder containing the raw
MEaSURES data; necessary for cropping for example. Set as NULL if no
MEaSURES data is necessary / available; if so, any functions with a
spatial aspect will not be available (e.g. "crop_racmo_to_x",
"draw_antarctica"). It may be necessary to manually amend the folder and
file names for these (starting line 79 in the raw code of this function).}
}
\description{
racmoR functions need to know where certain datasets are. This
function helps access these datasets by adding their paths to the global
environment (as an object called "racmoInfo"). This function should be
called at the start of a session. The datasets must already be saved and
accessible. As racmoR expands, this function could also set different
variables that are necessary between projects.
}
