% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_racmo_axes.R
\name{draw_racmo_axes}
\alias{draw_racmo_axes}
\title{Add a tighter border of axes around Antarctic RACMO plots}
\usage{
draw_racmo_axes(
  x = "",
  crs = "racmo",
  tickKula = "#999999FF",
  axisKula = "#999999FF",
  interval = NULL,
  tickLength = -0.25
)
}
\arguments{
\item{x}{SpatRaster: Which data has been plotted? Supply the SpatRaster to
correctly align the axes with the extent; if an empty string is supplied,
which is the default, it assumes that the whole continent has been
plotted, and will add the appropriate axes.}

\item{crs}{Which crs are the axes being added to?}

\item{tickKula}{What colour should the tick marks be?}

\item{axisKula}{What colour should the axis lines be?}

\item{interval}{numeric: The interval between tick marks, in the ?}

\item{tickLength}{numeric: How long should the tick marks be?}
}
\description{
The default terra borders for Antarctica leave a lot of white
space around the continent. This function draws axes that are tighter to
the coastline of the raster. This function should be called straight
after using \code{terra::plot(x, axes = FALSE)}.
}
