% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coastlines.R
\name{get_coastline}
\alias{get_coastline}
\title{Return the MEaSURES coastline for a given extent}
\usage{
get_coastline(extent = "", crs = "racmo", preferType = NULL, crsIn = NULL)
}
\arguments{
\item{extent}{Define the area within which to get the coastline. This
argument is fed into \code{get_extent()}; see there for details.}

\item{crs}{"string": Which projection should the coastline be returned in?}

\item{preferType}{"string": Use ice shelves or basins if there is a clash?
See \code{get_extent()} for a full explanation.}

\item{crsIn}{"string": Which projection is the extent given in? Needs
defining if the extent is a SpatExtent, as they do not have crs value
attached.}
}
\description{
This function is useful for plotting the Antarctic coastline
(i.e. the most seaward part of the ice, grounded or shelf). It is mainly
used in the \code{draw_antarctica()} function, but works separately as well.
The coastlines can be reprojected using the "crs" argument, defaulting to
"racmo".
}
\examples{
  # Full coastline
  t1 <- get_coastline()
  terra::plot(t1)

  # Single shelf
  t2 <- get_coastline("Amery")
  terra::plot(t2)

  # Multiple shelves
  t3 <- get_coastline(c("Amery", "West"))
  terra::plot(t3)

  # Compare preferType options
  t4 <- get_coastline("Drygalski", preferType = NULL)
  terra::plot(t4)

  t5 <- get_coastline("Drygalski", preferType = "shelves")
  terra::plot(t5)

  t6 <- get_coastline("Drygalski", preferType = "basins")
  terra::plot(t6)

}
