% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent.R
\name{get_extent}
\alias{get_extent}
\title{Return the extent of different Antarctic boundaries}
\usage{
get_extent(
  extent = "",
  rectangularExtent = FALSE,
  preferType = NULL,
  returnOnly = NULL,
  crs = "racmo",
  crsIn = NULL
)
}
\arguments{
\item{extent}{Multiple types are allowed as input.

If extent is an empty string (i.e. "", the default), the whole of
Antarctica is returned.

If extent is a defined string (e.g. "Shackleton"), basins and ice shelves
in the MEaSURES dataset with those names are returned; if the same name
refers to both an ice shelf and a basin, 'preferType' determines which is
returned.

A vector of multiple ice shelf/basin names can also be used at once; if
that vector contains both ice shelves and basins but only one type is
required, use the 'returnOnly' argument.

If a defined string, by default the exact outline of the ice shelf or
basin is returned, but 'rectangularExtent' can be set for a bounding box
instead.

If extent is a SpatRaster or SpatVector, then the extent of the
SpatRaster or SpatVector will be returned.

If extent is a SpatExtent, it is necessary to also set 'crsIn', and a
SpatExtent will be returned.}

\item{preferType}{Sometimes an ice shelf and a basin have the same name;
usually they both refer to the same general area, and by default (i.e.
preferType = NULL), the two extents are combined to form a new larger
extent encompassing both. However, for some names (e.g. "Shackleton" and
"Drygalski"), the name refers to a shelf far away from the basin with the
same name.

This argument lets the user choose between the two types if the name is
found in both the ice shelves and basins datasets: "shelf", "floating" or
"FL" to return the ice shelves; "basin", "grounded" or "GR" to return the
basins. All other names that aren't in both the ice shelves and basins
datasets will be returned regardless of this option; see also returnOnly.}

\item{returnOnly}{If there is a list containing both ice shelf and basin
names, but only one or the other should be returned, use this argument.
Ice shelves or basins are specified as outlined in the preferType
argument.}

\item{crs}{"string": Which crs should the extent be returned with? See
\code{use_crs()} or \code{terra::crs()}.}

\item{crsIn}{"string": Which projection is the extent given in? Needs
defining if the extent is a SpatExtent, as they do not have crs value
attached.}

\item{rectangulerExtent}{If extent is a defined string (e.g. "Amery"), the
rectangulerExtent argument determines whether the actual outline is
returned (FALSE; the default), or if a rectangular extent box
encompassing the area is returned (TRUE). If multiple ice shelves and/or
basins are included, the extent box will extend over all of them at once.}
}
\description{
This function helps us to very easily crop and/or plot data
for parts of Antarctica. It is very flexible in what it accepts, but is
particularly useful when entering MEaSURES ice shelf or basin names.
However, it can also handle other SpatRasters, SpatVectors, or
SpatExtents, allowing it to be used in different contexts. The extent can
be returned in different projections via the crs argument. The returned
extent can be either exact shape outlines or a rectangular SpatExtent.
}
