% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grounding_line.R
\name{get_grounding_line}
\alias{get_grounding_line}
\title{Return the MEaSURES grounding line for a given extent}
\usage{
get_grounding_line(extent = "", crs = "racmo", preferType = NULL, crsIn = NULL)
}
\arguments{
\item{extent}{Define the area within which to get the grounding line. This
argument is fed into \code{get_extent()}; see there for details.}

\item{crs}{"string": Which projection should the grounding line be returned
in?}

\item{preferType}{"string": Use ice shelves or basins if there is a clash
in the extent definition? See \code{get_extent()} for a full explanation.}

\item{crsIn}{"string": Which projection is the extent given in? Needs
defining if the extent is a SpatExtent, as they do not have crs value
attached.}
}
\description{
This function is useful for plotting the grounding line (GL)
of Antarctica. It is mainly used in the \code{draw_antarctica()} function, but
works separately as well. The GL can be reprojected using the "crs"
argument, defaulting to "racmo".
}
\examples{
  # Full grounding line
  t1 <- get_grounding_line()
  terra::plot(t1)

  # Single shelf
  t2 <- get_grounding_line("Amery")
  terra::plot(t2)

  # Multiple shelves
  t3 <- get_grounding_line(c("Amery", "West"))
  terra::plot(t3)

  # Compare preferType options
  t4 <- get_grounding_line("Drygalski", preferType = NULL)
  terra::plot(t4)

  t5 <- get_grounding_line("Drygalski", preferType = "shelves")
  terra::plot(t5)

  t6 <- get_grounding_line("Drygalski", preferType = "basins")
  terra::plot(t6)

}
