% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_racmoD_data.R
\name{get_racmoD_file_path}
\alias{get_racmoD_file_path}
\title{Create the file path for the daily racmo data}
\usage{
get_racmoD_file_path(variable, year)
}
\arguments{
\item{variable}{"string": Which RACMO variable?}

\item{year}{numeric: Which year is the data from?}
}
\description{
The daily RACMO data is split across NetCDF files by decade.
This function simply retrieves the filepaths of the different NetCDFs.
Only used by \code{read_racmoD_data()}.
}
