% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shelf_outline.R
\name{get_shelf_outline}
\alias{get_shelf_outline}
\title{Return the MEaSURES shelf outlines for a given extent}
\usage{
get_shelf_outline(
  extent = "",
  exactExtents = TRUE,
  crs = "racmo",
  crsIn = NULL
)
}
\arguments{
\item{extent}{Define the area within which to get the shelf outlines. Exact
shelf names can be used, as can SpatRasters, SpatVectors, or SpatExtents,
or it can be left empty, and all shelves are returned. For exactExtents =
TRUE to work, this argument must be a string with the MEaSURES names of
the required shelves. Fed into \code{get_extent()}; see there for details.}

\item{exactExtents}{BINARY: Should only the defined shelf/shelves be
included (TRUE), or can the outlines of all shelves within the bounding
box be included too (FALSE)?}

\item{crs}{"string": Which projection should the grounding line be returned
in?}

\item{crsIn}{"string": Which projection is the extent given in? Needs
defining if the extent is a SpatExtent, as they do not have crs value
attached.}
}
\description{
This function is useful for plotting ice shelves in
Antarctica. It is mainly used in the \code{draw_antarctica()} function, but
works separately as well. The ice shelves can be reprojected using the
"crs" argument, defaulting to "racmo".
}
\examples{
  # All ice shelves in Antarctica
  t1 <- get_shelf_outline()
  terra::plot(t1)

  # Just on shelf  t2 <- get_shelf_outline("Amery")
  terra::plot(t2)

  # Multiple shelves; only the named shelves
  t3 <- get_shelf_outline(c("Amery", "West"), exactExtents = TRUE)
  terra::plot(t3)

  # Multiple shelves; includes intermediary shelves
  t4 <- get_shelf_outline(c("Amery", "West"), exactExtents = FALSE)
  terra::plot(t4)

}
