% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_variables.R
\name{list_marD_variables}
\alias{list_marD_variables}
\title{Return the available daily MAR variables}
\usage{
list_marD_variables(print = TRUE)
}
\arguments{
\item{print}{BINARY: Should the variables be printed out? Can be suppressed
for use inside other functions.}
}
\description{
Which daily MAR variables are available? It simply reads the
names of the files as created by \code{\link[=configure_polarcm]{configure_polarcm()}} (and thus only
accounts for the files present when that function was initially called).
}
