% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_racmo_variables.R
\name{list_racmoM_variables}
\alias{list_racmoM_variables}
\title{Return the available monthly RACMO variables}
\usage{
list_racmoM_variables(print = TRUE)
}
\arguments{
\item{print}{BINARY: Should the variables be printed out? Can be suppressed
for use inside other functions.}
}
\description{
Which monthly RACMO variables are available? It simply reads
the names of the files as created by \code{\link[=configure_polaR]{configure_polaR()}} (and thus
only accounts for the files present when that function was initially
called).
}
