% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_racmo.R
\name{mask_racmo_to_basin}
\alias{mask_racmo_to_basin}
\title{Mask RACMO data to an Antarctic basin}
\usage{
mask_racmo_to_basin(racmoData, basin)
}
\arguments{
\item{racmoData}{SpatRaster: The RACMO data to mask. It must be an existing
SpatRaster.}

\item{basin}{"string": The EXACT name of the basins to mask the RACMO data
to. If multiple are used, non-basins pixels in the bounding box are also
masked out.}
}
\description{
Spatially mask RACMO SpatRasters. 'basin' must be an exact
match for a name in the MEaSURES or IMBIE dataset. Multiple basins can be
used in a vector. The data returned is for the actual outline around the
basins; use \code{crop_racmo_to_basin()} for a rectangular bounding box.
}
\examples{
-----------------------------------------------------------------
  dontrun{
    x <- subset_racmoM_by_summer("precip", 1991) |>
      subset_racmoM_by_month(12)

    # Single basin
    xMask <- mask_racmo_to_basin(x, "A-Ap")
    terra::plot(xMask)

    # Multiple basins
    xMask2 <- mask_racmo_to_basin(x, c("A-Ap", "Dry Valleys"))
    terra::plot(xMask2)
  }

}
\seealso{
crop_racmo_to_basin()

mask_racmo_to_shelf()
}
