% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_racmo.R
\name{mask_racmo_to_shelf}
\alias{mask_racmo_to_shelf}
\title{Mask RACMO data to an Antarctic ice shelf}
\usage{
mask_racmo_to_shelf(racmoData, shelf)
}
\arguments{
\item{racmoData}{SpatRaster: The RACMO data to mask. It must be an existing
SpatRaster.}

\item{shelf}{"string": The EXACT name of the ice shelves to mask the
RACMO data to. If multiple are used, non-shelf pixels in the bounding box
are masked out.}
}
\description{
Spatially mask RACMO SpatRasters. 'shelf' must be an exact
match for a name in the MEaSURES dataset. Multiple shelves can be used in
a vector. The data returned is for the actual outline around the shelves;
use \code{crop_racmo_to_shelf()} for a rectangular bounding box.
}
\examples{
-----------------------------------------------------------------
  dontrun{
    x <- subset_racmoM_by_summer("precip", 1991) |>
      subset_racmoM_by_month(12)

    # Single shelf
    xMask <- mask_racmo_to_shelf(x, "Amery")
    terra::plot(xMask)

    # Multiple shelves
    xMask2 <- mask_racmo_to_shelf(x, c("Amery", "Shackleton"))
    terra::plot(xMask2)
  }

}
\seealso{
crop_racmo_to_shelf()

mask_racmo_to_basin()
}
