% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_to_x.R
\name{mask_to_basin}
\alias{mask_to_basin}
\title{Mask RCM data to an Antarctic basin}
\usage{
mask_to_basin(x, basin)
}
\arguments{
\item{x}{SpatRaster: The RCM data to mask. It must be an existing
SpatRaster.}

\item{basin}{"string": The EXACT name of the basins to mask 'x' to. If
multiple are used, non-basins pixels in the bounding box are also masked
out.}
}
\description{
Spatially mask RCM SpatRasters. The 'basin' must be an exact
match for a basin name in the MEaSURES dataset, and can refer to either
the IMBIE basins (e.g. "A-Ap") or the refined MEaSURES basins (e.g. "Dry
Valleys"). Multiple basins can be used in a vector.

The data returned is "masked" to the actual basin outline, meaning the
surrounding pixels are given a value of NA; use \code{\link[=crop_to_basin]{crop_to_basin()}} for the
rectangular bounding box.
}
\examples{
-----------------------------------------------------------------
\dontrun{
  x <- subset_racmoM_by_summer("precip", 1991) |>
    subset_racmoM_by_month(12)

  # Single basin
  xMask <- mask_to_basin(x, "A-Ap")
  terra::plot(xMask)

  # Multiple basins
  xMask2 <- mask_to_basin(x, c("A-Ap", "Dry Valleys"))
  terra::plot(xMask2)
}

}
\seealso{
crop_to_basin()

mask_to_shelf()
}
