% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_racmoD.R
\name{read_racmoD}
\alias{read_racmoD}
\title{Read in daily RACMO data from file}
\usage{
read_racmoD(racmoData, version = NULL)
}
\arguments{
\item{racmoData}{The daily RACMO data to return. If a valid string, raw
daily RACMO data for the corresponding variable is read in from the
respective NetCDF files. Use \code{\link[=list_racmoD_variables]{list_racmoD_variables()}} to see the valid
variable names. If an existing SpatRaster, it is simply returned as it
is.}

\item{version}{Which version of the daily RACMO data should be used? The
options are those defined by the user in their ".Rprofile" file; see
\code{\link[=configure_polarcm]{configure_polarcm()}} for more information. By default, it uses the
racmoD data that was defined first in the ".Rprofile". Ignored if
'racmoData' is already a SpatRaster.}
}
\description{
Read in the NetCDF files for daily resolution RACMO data
(racmoD). This function returns a SpatRaster (see the \code{terra} package),
and can handle the necessary configurations for the racmoD data.
}
