% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_racmoD_data.R
\name{read_racmoD_data}
\alias{read_racmoD_data}
\title{Read in daily RACMO data from file}
\usage{
read_racmoD_data(racmoData)
}
\arguments{
\item{variable}{"string": If a valid string is input (i.e. it must match an
option returned by \code{list_racmoD_variables()}), the daily RACMO data for
that variable is read from the corresponding NetCDF files. If a
SpatRaster is entered, it is simply returned, allowing this function to
be used flexibly.}
}
\description{
This function allows other functions to easily access monthly
RACMO data. It will return a SpatRaster, either by reading the necessary
NetCDF file, or by simply returning the SpatRaster it is given. It is the
daily equivalent of \code{read_racmoM_data()}, but works a little differently
underneath because it assumes that the daily data has been separated into
decadal NetCDFs.
}
