% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_racmoM.R
\name{read_racmoM}
\alias{read_racmoM}
\title{Read in monthly RACMO data from file}
\usage{
read_racmoM(racmoData, version = NULL)
}
\arguments{
\item{racmoData}{The monthly RACMO data to return. If a valid string, raw
monthly RACMO data for the corresponding variable is read in from the
respective NetCDF file. Use \code{\link[=list_racmoM_variables]{list_racmoM_variables()}} to see the valid
variable names. If an existing SpatRaster, it is simply returned as it
is.}

\item{version}{Which version of the monthly RACMO data should be used? The
options are those defined by the user in their ".Rprofile" file; see
\code{\link[=configure_polarcm]{configure_polarcm()}} for more information. By default, it uses the
racmoM data that was defined first in the ".Rprofile". Ignored if
'racmoData' is already a SpatRaster.}
}
\description{
Read in the NetCDF files for monthly resolution RACMO data
(racmoM). This function returns a SpatRaster (see the \code{terra} package),
and knows how to handle some different configurations of the racmoM
\code{polaR} functions need.
}
