% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_racmoM_data.R
\name{read_racmoM_data}
\alias{read_racmoM_data}
\title{Read in monthly RACMO data from file}
\usage{
read_racmoM_data(variable)
}
\arguments{
\item{variable}{"string": If a valid string is input (i.e. it must match an
option returned by \code{list_racmoM_variables()}), the monthly RACMO data for
that variable is read from the corresponding NetCDF file. If a SpatRaster
is entered, it is simply returned, allowing this function to be used
flexibly.}
}
\description{
This function allows other functions to easily access monthly
RACMO data. It will return a SpatRaster, either by reading the necessary
NetCDF file, or by simply returning the SpatRaster it is given.

The RACMO data must already be stored on the machine, and the path to
defined using \code{define_racmo_globals()}. The SpatRaster's crs and extent
are explicitly hardcoded as \code{terra} fails to read them from the file
correctly. They are based on RACMO2.3p3; adjust if the RACMO version data
differs.
}
