% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_racmoD_by_date.R
\name{subset_racmoD_by_date}
\alias{subset_racmoD_by_date}
\title{Subset daily RACMO data based on specific dates}
\usage{
subset_racmoD_by_date(racmoData, dates = NULL, dateList = NULL, version = NULL)
}
\arguments{
\item{racmoData}{The daily RACMO data to subset. Can be either a variable
name, in which case raw daily RACMO data is read in; or an existing
SpatRaster of daily RACMO data.}

\item{dates}{A vector of specific dates to return. Each date must be in the
format "YYYY-MM-DD".}

\item{dateList}{A list of vectors, with the first value in each vector
indicating when the period begins, and the second when the period ends.
Each date must be in the format "YYYY-MM-DD".}

\item{version}{Which version of the daily RACMO data should be returned?
The options are those defined by the user in their ".Rprofile" file; see
\code{\link[=configure_polaR]{configure_polaR()}} for more information. By default, it uses the racmoD
data that was defined first in the ".Rprofile". Ignored if 'racmoData' is
already a SpatRaster.}
}
\description{
Subset daily RACMO data based on specific dates. This function
is essentially a RACMO-specific wrapper around
\code{terrapin::subset_by_dates()}. The dates can be input either as a vector
of specific dates using the "dates" argument; or as a list of vectors
containing the start and end dates for longer extended periods. All dates
need to be entered in YYYY-MM-DD format (e.g. "2019-12-31").
}
