% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_racmoM_by_month.R
\name{subset_racmoM_by_month}
\alias{subset_racmoM_by_month}
\title{Subset monthly RACMO data based on the month}
\usage{
subset_racmoM_by_month(racmoData, months, removeIncomplete = 3)
}
\arguments{
\item{months}{vector: Which month/s to return? Input can be the month
number (e.g. 12) or the month name, either in full ("December",
"december") or abbreviated ("Dec", "dec"). Multiple months can be input
at once (e.g. c(12, 1, 2)), but do not try to mix strings and numbers in
the vector.}

\item{removeIncomplete}{If the value is "years", the data is run through
\code{terrapin::remove_incomplete_years()}, and only months in years with all
requested months in are returned. If the value is numeric (between 1 and
12), the data is fed into the
\code{terrapin::remove_incomplete_austral_summers()} and the value is used as
the "australSplit" argument to return only months in austral summers that
contain all requested months. If any other value, all layers matching the
months argument are returned, regardless of the summer or year. See the
examples.}

\item{racmoData:}{The monthly RACMO data to subset. Can be either a
variable name, in which case raw monthly RACMO data is read in; or an
existing spatRaster of monthly RACMO data.}
}
\description{
Subset monthly RACMO data based on month. This function is
essentially a RACMO-specific wrapper around
\code{terrapin::subset_by_month()}.
}
