% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_racmoM_by_summer.R
\name{subset_racmoM_by_summer}
\alias{subset_racmoM_by_summer}
\title{Subset monthly RACMO data based on the austral summer}
\usage{
subset_racmoM_by_summer(
  racmoData,
  summers = NULL,
  australSplit = 3,
  after = NULL,
  before = NULL,
  except = NULL,
  version = NULL
)
}
\arguments{
\item{racmoData}{The monthly RACMO data to subset. Can be either a variable
name, in which case raw monthly RACMO data is read in; or an existing
SpatRaster of monthly RACMO data.}

\item{summers}{Which summer/s to return? Use this argument for exact
matches (e.g. c(1980:1990, 1995:1997). Cannot be used in conjunction with
the 'before', 'after' or 'except' arguments.}

\item{australSplit}{numeric: If not FALSE, create an additional column for
the austral summer / year. For example, December 1991 and January 1992
often need to be considered as part of summer 1991/1992 in the southern
hemisphere rather than in their respective years.

The default value is 3, which means that all months \emph{AFTER} March (i.e.
month 3) are considered as part of the following austral summer / year
(i.e. April 1991 -- March 1992 are all in austral summer / year 1992).

Swap this value accordingly: e.g. setting it as 4 means May 1991 -- April
1992 are all austral summer / year 1992.}

\item{after}{Only data that occurs after this value is returned. Cannot be
used in conjunction with the exact values, 'before' or 'except'
arguments.}

\item{before}{Only data that occurs before this value is returned. Cannot
be used in conjunction with the exact values, 'after' or 'except'
arguments.}

\item{except}{vector: Which data should be excluded from the SpatRaster?
Cannot be used in conjunction with the exact values, 'after' or 'before'
arguments.}

\item{version}{Which version of the monthly RACMO data should be used? The
options are those defined by the user in their ".Rprofile" file; see
\code{\link[=configure_polarcm]{configure_polarcm()}} for more information. By default, it uses the
racmoM data that was defined first in the ".Rprofile". Ignored if
'racmoData' is already a SpatRaster.}
}
\description{
Subset monthly RACMO data based on the austral summer. An
austral summer is defined using the "australSplit" argument; see that for
an explanation. This function is essentially a wrapper around
\code{\link[terrapin:subset_by_summer]{terrapin::subset_by_summer()}} for monthly RACMO data.
}
