% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_racmoM_by_year.R
\name{subset_racmoM_by_year}
\alias{subset_racmoM_by_year}
\title{Subset monthly RACMO data based on the year}
\usage{
subset_racmoM_by_year(
  racmoData,
  years = NULL,
  after = NULL,
  before = NULL,
  except = NULL,
  version = NULL
)
}
\arguments{
\item{racmoData}{The monthly RACMO data to subset. Can be either a variable
name, in which case raw monthly RACMO data is read in; or an existing
SpatRaster of monthly RACMO data.}

\item{years}{Which year/s to return? Use this argument for exact matches
(e.g. c(1983:1986, 1991:1992). Cannot be used in conjunction with
the 'before', 'after' or 'except' arguments.}

\item{after}{Only data that occurs after this value is returned. Cannot be
used in conjunction with the exact values, 'before' or 'except'
arguments.}

\item{before}{Only data that occurs before this value is returned. Cannot
be used in conjunction with the exact values, 'after' or 'except'
arguments.}

\item{except}{vector: Which data should be excluded from the SpatRaster?
Cannot be used in conjunction with the exact values, 'after' or 'before'
arguments.}

\item{version}{Which version of the monthly RACMO data should be used? The
options are those defined by the user in their ".Rprofile" file; see
\code{\link[=configure_polaR]{configure_polaR()}} for more information. By default, it uses the racmoM
data that was defined first in the ".Rprofile". Ignored if 'racmoData' is
already a SpatRaster.}
}
\description{
Subset monthly RACMO data using based on year. This function
is essentially just a wrapper around \code{terrapin::subset_by_year()} for
monthly RACMO data.
}
