% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_racmo_name.R
\name{use_racmo_name}
\alias{use_racmo_name}
\title{Help with labelling of RACMO variables and units}
\usage{
use_racmo_name(
  racmoVar,
  originalUnits = TRUE,
  monthlyData = FALSE,
  shortName = TRUE
)
}
\arguments{
\item{racmoVar}{Which RACMO variable to use? Use the short name code.}

\item{originalUnits}{Should the same units as the dataset be used (TRUE;
default)? Often this makes the most sense, but some of the variables are
often changed in the same way for ease of use. For example, it is usually
easier to think of snow/ice temperatures in ºC rather than Kelvin, to use
Wm-2 rather than total joules, and to think of melt or precipitation as
daily totals rather than fluxes per second. Be \emph{VERY VERY VERY} careful
if setting this as FALSE - it just assumes that the changes are the
common ones that I have previously done, there is nothing clever about
it!}

\item{monthlyData}{Is the RACMO data monthly resolution (i.e. racmoM; TRUE)
or daily (i.e. racmoD; FALSE, the default)? For some variables, this
makes a difference (e.g. all radiative fluxes are in Joules in monthly
data, but W m-2 in monthly data); for others (e.g. "t2m", "u10m") it
makes no difference. Set as NULL to suppress the unit output, and just
return the variable name.}

\item{shortName}{Should shortened names ("Precip.") or acronyms ("SMB") be
used over the full name (e.g. "Precipitation", "Surface Mass Balance")?
If set as NULL, no variable name is used, and only the units are
returned.}
}
\description{
"senf" isn't clear on a plot, it should be (for example)
"Sensible Heat Flux (W m-2)". However, that requires a lot of repeated
typing, particularly when concerned with the superscripts and subscripts
with \code{bquote}. This function is (a work-in-progress) way to automate
this.
}
