% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fun_in_timeM.R
\name{calc_fun_annualM}
\alias{calc_fun_annualM}
\title{Calculate RCM values over multiple months each year}
\usage{
calc_fun_annualM(x, months, FUN, ...)
}
\arguments{
\item{x}{SpatRaster: The RACMO or MAR data to use with 'FUN'. It must be
an existing SpatRaster.}

\item{months}{vector: Which month/s to include? Input can be the month
number (e.g. 12) or the month name, either in full ("December",
"december") or abbreviated ("Dec", "dec"). Multiple months can be input
at once (e.g. c(12, 1, 2)), but do not try to mix strings and numbers in
the vector.}

\item{FUN}{Which function should be applied? Examples include "mean", "sd",
and "median". This function is applied to each pixel across the months
each year. For example, "mean" gives each pixel's average value across
the months each year; using "sum" would calculate the total across those
months each year.}

\item{...}{Any arguments that should be passed to the 'FUN' function.}
}
\description{
Often, it is necessary to look at RCM values across multiple
months each year. For example, what is the total JJA precipitation each
year? Or what is the average wind speed in SON? This function takes the
input RCM data, and creates new SpatRasters based on the requested
'months' and the 'FUN' (function) arguments.

The following example further illustrates this function. Take a monthly
RACMO dataset of precipitation, including all months each year from
1979--2018 (i.e. 12 months * 40 years = 480 layers). Set 'months' as c(6,
7, 8) (i.e. JJA), and FUN as "sum". The returned SpatRaster will contain
40 layers, with each one being that year's total JJA precipitation. The
date of each returned layer is date of the first of the constituent
layers that it was computed from. Only layers with the same units are
processed together, and only years when all requested months are
available are returned.
}
\seealso{
calc_fun_australM
}
