% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_racmoD_climatology.R
\name{calc_racmoD_climatology}
\alias{calc_racmoD_climatology}
\title{Apply a function to daily RACMO data across multiple years}
\usage{
calc_racmoD_climatology(racmoData, FUN, excludeIncomplete = FALSE, ...)
}
\arguments{
\item{racmoData}{SpatRaster: The RACMO data. It needs to be an existing
SpatRaster.}

\item{FUN}{Which function should be applied across the years? For example,
if set as "mean", this function will return the mean value of the
racmoData variable on the 1st Dec each year. Can be a vector of
functions, but be wary of speed and there is no elegant solution to
supplying ... for different functions; it is probably easier to run this
separately for each required function and then combine them as a
SpatRasterDataset.}

\item{excludeIncomplete}{Should data be excluded if it is not found in
throughout the racmoData dataset? TRUE removes dates that not found in
every calendar year (i.e. Jan--Dec); a numeric value between 1 and 12 is
used as the australSplit (i.e. if 3, all months AFTER March, are
considered part of the following summer). If FALSE, all data is used when
calculating the function.

Can be useful so the function is applied to the same years/summers for
each day, but be careful! Worst case scenario: if there were 5 years of
complete data (e.g. January 1 2000 - December 31 2005), and a single day
outside this (e.g. January 1 2006), the function would only be applied to
the January 1st data as the other dates are not found in 2006. See
\code{exclude_unmatched_days}. It is possibly much easier and clearer to
establish which dates are in the dataset first, and then run this
function with 'excludeIncomplete' set as FALSE.}

\item{...}{Any arguments that need to be fed to FUN. If there are multiple
values in a FUN vector, these arguments will be fed to all of the
functions.}
}
\description{
What is the mean air temperature on the 1st December each
summer? What is the maximum wind speed on the days from the 10th to the
15th July each year? What is the standard deviation in surface melt on
the days in the first weeks of January? This function can help answer
such questions.
}
