% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_racmoR.R
\name{configure_racmoR}
\alias{configure_racmoR}
\title{Define paths to RACMO datasets and variables}
\usage{
configure_racmoR()
}
\description{
Please read the Instructions below!

\code{racmoR} functions need to know where certain datasets are on your system
to work - the \code{configure_racmoR()} function helps functions access these
datasets by creating the necessary paths and adding them to a hidden
".racmoR" environment. This function therefore needs to be called before
any racmoR functions are used (which is automatically done as part of
most \code{racmoR} functions). The datasets must already be saved and
accessible on your system.
}
\section{Instructions}{
This function must be called before using any \code{racmoR} functions (once
per R session). The raw data paths need to be defined in an ".Rprofile"
file. The ".Rprofile" file should be in either: the current working
directory (\code{\link[=getwd]{getwd()}}), the user's home directory (\code{Sys.getenv("HOME")}),
or the R installation (\code{\link[=R.home]{R.home()}}). The ".Rprofile" file is simply a file
called ".Rprofile" that runs automatically when R starts a new session.
For more info on ".Rprofile" files, see Section 2.4 of
\url{https://csgillespie.github.io/efficientR/set-up.html}.

The following code should be pasted into the ".Rprofile" file. This will
let \code{racmoR} where to find the RACMO data. Therefore, you must adjust the
paths to match your directory structure; the paths should be relative to
your current working directory.

\if{html}{\out{<div class="sourceCode R">}}\preformatted{  # Prepare data paths for racmoR
  .racmoR = new.env() # hidden racmoR environment
  .racmoR$rawDataPath <- "../../Data/
  .racmoR$racmoM      <- "RACMO/RACMO2.3p3_CON_ANT27_monthly/"
  .racmoR$racmoD      <- "RACMO/RACMO2.3p3_CON_ANT27_daily/"
  .racmoR$MEaSURES    <- "MEaSURES Boundaries/"
  attach(.racmoR)
}\if{html}{\out{</div>}}

This function assumes that all raw data is stored in a "Data" directory
(i.e. 'rawDataPath'), with separate subdirectories for RACMO data at
monthly ('racmoM') and daily ('racmoD') resolution, and for the
'MEaSURES' data. The necessary data must already be downloaded before
using \code{racmoR} - the package \emph{does not} try to find and download data for
you. If any datasets are not available, set their path as \code{NULL}. NULLs
will restrict the capabilities of the racmoR package in expected ways -
for example, if no daily RACMO data is available, the racmoD functions
will fail. However, that may not be an issue. The MEaSURES data is used
for any location-based calculations or subsetting, and for drawing
borders.
}

