% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_antarctica.R
\name{draw_antarctica}
\alias{draw_antarctica}
\title{Draw MEaSURES Antarctic outlines on maps}
\usage{
draw_antarctica(
  extent = "",
  sbcg = "cg",
  exactExtents = TRUE,
  simplify = 0,
  crs = "racmo",
  ...
)
}
\arguments{
\item{extent}{Define the area within which to draw Antarctica. This
argument is fed into \code{get_extent()}; see there for details.}

\item{sbcg}{"string": Which parts of Antarctica should be drawn? Options
are:
\itemize{
\item shelves          "s"
\item basins           "b"  # uses IMBIE; use "bb" for MEaSURES
\item coasts           "c"
\item grounding line   "g"
}}

\item{exactExtents}{BINARY: Should only the defined shelf/shelves be
included (TRUE), or can the outlines of all shelves within the bounding
box be included too (FALSE)?}

\item{simplify}{numeric: Should the outline be simplified? Uses the
\code{terra::simplifyGeom()} function, so this value is the tolerance - i.e.
nodes must be at least this far apart, defined in crs units. Larger
numbers are coarser; 0 (default) is no simplifying.}

\item{crs}{"string": Which crs should the lines be drawn in? See the
\code{set_crs()} function}

\item{...}{Any arguments that can be used in \code{terra::lines()}}
}
\description{
Used after plotting a raster to add MEaSURES outlines. This
function is essentialy a wrapper around \code{terra::lines()}, but lets us
easily decide on the outlines to add, their extent, and with which crs.
}
