% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basin_outline.R
\name{get_basin_outline}
\alias{get_basin_outline}
\title{Return the MEaSURES basin outlines for a given extent}
\usage{
get_basin_outline(
  extent = "",
  exactExtents = TRUE,
  subBasins = FALSE,
  crs = "racmo",
  crsIn = NULL
)
}
\arguments{
\item{extent}{Define the area within which to get the basin outlines. Exact
shelf names can be used, as can SpatRasters, SpatVectors, or SpatExtents,
or it can be left empty, and all basins are returned. For exactExtents =
TRUE to work, this argument must be a string with the MEaSURES names of
the required shelves. Fed into \code{get_extent()}; see there for details.}

\item{exactExtents}{BINARY: Should only the defined shelf/shelves be
included (TRUE), or can the outlines of all shelves within the bounding
box be included too (FALSE)?}

\item{subBasins}{BINARY: Should the MEaSURES subregion basins be returned?
Default is FALSE, which means that the IMBIE basins are returned. If
exactExtents is TRUE, make sure that the extent argument uses the correct
name for the correct dataset: IMBIE basins (subBasins = FALSE) have codes
(e.g. "A-Ap"), whereas MEaSURES basins (subBasins = TRUE) have names,
such as "Vincennes_Bay".}

\item{crs}{"string": Which projection should the grounding line be returned
in?}

\item{crsIn}{"string": Which projection is the extent given in? Needs
defining if the extent is a SpatExtent, as they do not have crs value
attached.}
}
\description{
This function is useful for plotting basins in Antarctica. It
is mainly used in the \code{draw_antarctica()} function, but works separately
as well. The basins can be reprojected using the "crs" argument,
defaulting to "racmo".
}
