% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basin_outline.R
\name{get_basin_outline}
\alias{get_basin_outline}
\title{Return the MEaSURES basin outlines for a given extent}
\usage{
get_basin_outline(
  extent = "",
  exactExtents = TRUE,
  imbieBasins = NULL,
  crs = "racmo",
  crsIn = NULL
)
}
\arguments{
\item{extent}{Define the area within which to get the basin outlines. Exact
basin names can be used, as can SpatRasters, SpatVectors, or SpatExtents,
or it can be left empty, and all basins are returned. For exactExtents =
TRUE to work, this argument must be a string with the MEaSURES names of
the required basins. Fed into \code{get_extent()}; see there for details.}

\item{exactExtents}{BINARY: Should only the defined basin/s be included
(TRUE), or can the outlines of all basins within the bounding box be
included too (FALSE)?}

\item{imbieBasins}{BINARY: Should the IMBIE basins (TRUE, the default) or
the MEaSURES basins (FALSE) be returned? The IMBIE basins have codes
(e.g. "A-Ap") and are larger regions; the MEaSURES basins have names
(that can overlap with the ice shelf names), and are small "sub" regions.
If NULL, both are returned.}

\item{crs}{"string": Which projection should the grounding line be returned
in?}

\item{crsIn}{"string": Which projection is the extent given in? Needs
defining if the extent is a SpatExtent, as they do not have crs value
attached.}
}
\description{
This function is useful for plotting basins in Antarctica. It
is mainly used in the \code{draw_antarctica()} function, but works separately
as well. The basins can be reprojected using the "crs" argument,
defaulting to "racmo".
}
