% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_racmoD.R
\name{read_racmoD}
\alias{read_racmoD}
\title{Read in daily RACMO data from file}
\usage{
read_racmoD(racmoData, version = NULL)
}
\arguments{
\item{racmoData}{If a valid string is input, the daily RACMO data for
that variable is returned as a SpatRaster from the corresponding NetCDF
file/s. Use \code{\link[=list_racmoD_variables]{list_racmoD_variables()}} to see the valid variables. If a
SpatRaster is entered, it is simply returned as it is.}

\item{version}{Which version of the daily RACMO data should be returned?
The options are those defined by the user in their ".Rprofile" file; see
\code{\link[=configure_polaR]{configure_polaR()}} for more information. By default, it uses the racmoD
data that was defined first in the ".Rprofile". Ignored if 'racmoData' is
already a SpatRaster.}
}
\description{
Read in the NetCDF files for daily resolution RACMO data
(racmoM). This function returns a SpatRaster (see the \code{terra} package),
but knows how to handle some different configurations of the racmoD
\code{polaR} functions need.
}
