% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_racmoD_by_day.R
\name{subset_racmoD_by_day}
\alias{subset_racmoD_by_day}
\title{Subset daily RACMO data based on the day of the month}
\usage{
subset_racmoD_by_day(racmoData, days)
}
\arguments{
\item{racmoData}{The daily RACMO data to subset. Can be either a variable
name, in which case raw daily RACMO data is read in; or an existing
SpatRaster of daily RACMO data.}

\item{days}{vector: Which day/s to return?}
}
\description{
Subset the daily RACMO data based on the day of the month. It
only works with the daily part of the date, so there is no accounting for
months; asking for 31 in February doesn't return any data layer. It will
return the data for multiple months (i.e. if 10 months are input as
"racmoData", setting "days = 1" will return the first day of each month).
}
