% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_racmoD_by_day.R
\name{subset_racmoD_by_day}
\alias{subset_racmoD_by_day}
\title{Subset daily RACMO data based on the day of the month}
\usage{
subset_racmoD_by_day(
  racmoData,
  days = NULL,
  before = NULL,
  after = NULL,
  except = NULL,
  version = NULL
)
}
\arguments{
\item{racmoData}{The daily RACMO data to subset. Can be either a variable
name, in which case raw daily RACMO data is read in; or an existing
SpatRaster of daily RACMO data.}

\item{days}{Which day/s to return? Use this argument for exact matches
(e.g. c(1, 8, 15:18). Cannot be used in conjunction with the 'before',
'after' or 'except' arguments.}

\item{before}{Only data that occurs before this value is returned. Cannot
be used in conjunction with the exact values, 'after' or 'except'
arguments.}

\item{after}{Only data that occurs after this value is returned. Cannot be
used in conjunction with the exact values, 'before' or 'except'
arguments.}

\item{except}{vector: Which data should be excluded from the SpatRaster?
Cannot be used in conjunction with the exact values, 'after' or 'before'
arguments.}

\item{version}{Which version of the daily RACMO data should be used? The
options are those defined by the user in their ".Rprofile" file; see
\code{\link[=configure_polarcm]{configure_polarcm()}} for more information. By default, it uses the
racmoD data that was defined first in the ".Rprofile". Ignored if
'racmoData' is already a SpatRaster.}
}
\description{
Subset daily RACMO data based on the day of the month. This
function is essentially a wrapper around \code{\link[terrapin:subset_by_day]{terrapin::subset_by_day()}} for
daily RACMO data.

\strong{Note:} This function is distinct to \code{\link[=subset_racmoD_by_date]{subset_racmoD_by_date()}} as it
accounts for only the day part of the date, not the full date (i.e. it
ignores the month and year).
}
