% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_racmoD_by_year.R
\name{subset_racmoD_by_year}
\alias{subset_racmoD_by_year}
\title{Subset daily RACMO data based on the year}
\usage{
subset_racmoD_by_year(racmoData, years, version = NULL)
}
\arguments{
\item{racmoData}{The daily RACMO data to subset. Can be either a variable
name, in which case raw daily RACMO data is read in; or an existing
SpatRaster of daily RACMO data.}

\item{years}{numeric: Which year/s to return?}
}
\description{
Subset daily RACMO data based on the year. This function is
essentially a RACMO-specific wrapper around \code{terrapin::subset_by_year()}.
}
