% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_racmoM_by_summer.R
\name{subset_racmoM_by_summer}
\alias{subset_racmoM_by_summer}
\title{Subset monthly RACMO data based on the austral summer}
\usage{
subset_racmoM_by_summer(racmoData, summers, australSplit = 3)
}
\arguments{
\item{summers}{numeric: Which summer/s to return?}

\item{australSplit}{numeric: Which is the last month included in an austral
summer before the new austral year begins? The default value is 3, which
means that all months \emph{AFTER} March are considered as part of the
following summer (i.e. April 1991 -- March 1992 are all in 1992). Swap
this value according: setting it as 4 means May 1991 -- April 1992 are
all 1992.}

\item{racmoData:}{The monthly RACMO data to subset. Can be either a
variable name, in which case raw monthly RACMO data is read in; or an
existing SpatRaster of monthly RACMO data.}
}
\description{
Subset monthly RACMO data based on the austral summer. This
function is essentially a RACMO-specific wrapper around
\code{terrapin::subset_by_summer()}. An austral summer is defined using the
"australSplit" argument; see that for an explanation.
}
