% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_crs.R
\name{use_crs}
\alias{use_crs}
\title{Return a crs string}
\usage{
use_crs(crs = NULL)
}
\arguments{
\item{crs}{"string": Which crs to return? See the "Options" section below
for a list of available projections. If none of those names are matched,
the string will be fed in as is, and assumed to be an alternative valid
crs definition. The default is to match whichever RCM data was listed
first in the ".Rprofile" file (match based on the dataset's 'src').}
}
\description{
RCM projections do not have standard EPSG codes, and usually
work on a rotated grid. This function returns the correct proj4 string
for the RACMO or MAR data in Antarctica. A couple of other common
projections for use in Antarctica are also available. These projection
definitions are stored in \code{.polarEnv$crs}, which is configured by
}
\section{Projections}{
\subsection{RACMO}{

The RACMO projections are taken directly from their respective NetCDFs.
These can be viewed with a command such as
\code{ncdf4::ncatt_get(nc, "rotated_pole")}.
}

\subsection{Projection Options}{

Options are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  - "racmo"             Antarctic wide
  - "stereo"            Antarctic Polar Stereographic (EPSG:3031)
  - "lambertEqual"      Lambert Equal Area Equidistant (EPSG: 102020)
  - "ortho"             Orthographic South Pole (EPSG: 102037)
  - "WGS84" / "ERA5"    WGS84 GPS System (EPSG: 4326)
  - "stereoAus"         Australian Antarctic Polar Stereographic (EPSG: 3032)
  - "lambertAus"        Australian Antarctic Lambert (EPSG: 3033)
  - "ease2"             NSIDC EASE-Grid 2.0 South (EPSG: 6932)
}\if{html}{\out{</div>}}

These crs definitions are stored in \code{.polarEnv$grids$crs}.
For more information on RCM projections, see here:
\url{https://gitlab.tudelft.nl/slhermitte/manuals/blob/master/RACMO_reproject.md}.

For more information on the non-RCM EPSG codes, visit a site such as
\url{https://epsg.io}, which is particularly good for showing the extent
where each projection is best suited for use. Add the EPSG code directly
to the end of that web address: \url{https://epsg.io/4326}.
}
}

